# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "This is an API for the Mechanic Shop"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
  /customers/login:
    post:
      tags:
        - Customers
      summary: "Login endpoint, (generates token)"
      description: "Validates customer credentials, generates and returns a customer specific token."
      parameters:
        - in: "body"
          name: "body"
          description: "Login credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Successful login"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              token: "eyJhbGciOiJI..."
              message: "Login successful"
              status: "success"
  /customers:
    post:
      tags:
        - Customers
      summary: "Endpoint to create a new customer."
      description: "Endpoint to create a new customer in our Mechanic Shop system, emails must be unique."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new customer."
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Customer created successfully"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "johnd@email.com"
              password: "123456"

    get:
      tags:
        - Customers
      summary: "Endpoint to retrieve all customers."
      description: "Endpoint to retrieve a list of all customers in our Mechanic Shop system."
      responses:
        200:
          description: "List of all customers retrieved successfully."
          schema:
            $ref: "#/definitions/AllCustomers"

    put:
      tags:
        - Customers
      summary: "Endpoint to update an existing customer's information."
      description: "Endpoint to update an existing customer's information in our Mechanic Shop system, this is a token authenticated route, and the user must be logged in to update their account."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to update an existing customer."
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Customer updated successfully"
          schema:
            $ref: "#/definitions/UpdateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: "Dohn Joe"
              email: "dohnj@email.com"
              password: "123456"

    delete:
      tags:
        - Customers
      summary: "Endpoint to delete an existing customer."
      description: "Endpoint to delete an existing customer from our Mechanic Shop system, this is a token authenticated route, and the user must be logged in to delete their account."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Customer deleted successfully"
          schema:
            $ref: "#/definitions/DeleteCustomerResponse"
          examples:
            application/json:
              message: "Customer <customer_id> deleted successfully"
  /customers/{customer_id}:
    get:
      tags:
        - Customers
      summary: "Endpoint to retrieve a customer by their ID."
      description: "Endpoint to retrieve a customer by their ID from our Mechanic Shop system."
      parameters:
        - in: "path"
          name: "customer_id"
          description: "ID of the customer to retrieve."
          required: true
          type: "integer"
      responses:
        200:
          description: "Customer retrieved successfully"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "johndoe@email.com"
              password: "123456"
  /customers/my-tickets:
    get:
      tags:
        - Customers
      summary: "Endpoint to retrieve all tickets for the logged in customer."
      description: "Endpoint to retrieve a list of all tickets for the logged in customer in our Mechanic Shop system, this is a token authenticated route, and the user must be logged in to view their tickets."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "List of all tickets for the logged in customer retrieved successfully."
          schema:
            $ref: "#/definitions/AllTicketsForCustomer"
...
definitions:
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
    
    required:
      - email
      - password
  LoginResponse:
    type: "object"
    properties:
      token:
        type: "string"
      message:
        type: "string"
      status:
        type: "string"
  CreateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
    
    required:
      - name
      - email
      - password
  CreateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
  AllCustomers:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        email:
          type: "string"
        password:
          type: "string"
  UpdateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
    
    required:
      - name
      - email
      - password
  UpdateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
  DeleteCustomerResponse:
    type: "object"
    properties:
      message:
        type: "string"
  AllTicketsForCustomer:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        customer_id:
          type: "integer"
        description:
          type: "string"
        status:
          type: "string"
...